moved to unsigned long for paranoid safety
git-svn-id: file:///source/coreland/corelib/trunk@562 6e7d7e42-f3d6-da11-80d8-00d0b73f4485

+++ UNIT_TESTS/t_alloc.c
@@ -9,7 +9,7 @@ static char *sp1;
 static char *sp2;
 static char *sp3;
 
-void set(char *p, unsigned int len)
+void set(char *p, unsigned long len)
 {
   for (;;) {
     if (!len) break; *p = 0x41; ++p; --len;
