+++ buffer_put.c
@@ -3,13 +3,13 @@
 #include "error.h"
 #include "str.h"
 
-static int write_all(int (*op)(int, char *, unsigned int),
-                     int fd, const char *s, unsigned int len)
+static int write_all(int (*op)(int, char *, unsigned long),
+                     int fd, const char *s, unsigned long len)
 {
   int w;
-  int (*cop)(int, const char *, unsigned int);
+  int (*cop)(int, const char *, unsigned long);
 
-  cop = (int (*)(int, const char *, unsigned int)) op;
+  cop = (int (*)(int, const char *, unsigned long)) op;
 
   while (len) {
     w = cop(fd, s, len);
@@ -33,9 +33,9 @@ int buffer_flush(buffer *b)
   return write_all(b->op, b->fd, b->buf, pos);
 }
 
-int buffer_put(buffer *b, const char *s, unsigned int len)
+int buffer_put(buffer *b, const char *s, unsigned long len)
 {
-  unsigned int n;
+  unsigned long n;
 
   n = b->size;
   if (len > n - b->pos) {
@@ -53,9 +53,9 @@ int buffer_put(buffer *b, const char *s, unsigned int len)
   return 0;
 }
 
-int buffer_putalign(buffer *b, const char *s, unsigned int len)
+int buffer_putalign(buffer *b, const char *s, unsigned long len)
 {
-  unsigned int n;
+  unsigned long n;
 
   while (len > (n = b->size - b->pos)) {
     bin_copy(s, b->buf + b->pos, n);
@@ -79,7 +79,7 @@ int buffer_putsalign(buffer *b, const char *s)
   return buffer_putalign(b, s, str_len(s));
 }
 
-int buffer_putflush(buffer *b, const char *s, unsigned int len)
+int buffer_putflush(buffer *b, const char *s, unsigned long len)
 {
   if (buffer_flush(b) == -1) return -1;
   return write_all(b->op, b->fd, s, len);
