+++ scan_space.c
@@ -1,6 +1,6 @@
 #include "scan.h"
 
-unsigned int scan_whitespace(const char *s)
+unsigned long scan_whitespace(const char *s)
 {
   return scan_charset(s, " \t");
 }
