+++ stalloc.h
@@ -4,8 +4,8 @@
 #define STALLOC_ALIGN 16
 #define STALLOC_SPACE 4096
 
-void *stalloc(unsigned int);
-int strealloc(void **, unsigned int, unsigned int);
+void *stalloc(unsigned long);
+int strealloc(void **, unsigned long, unsigned long);
 void stdealloc(void *);
 
 #endif
