+++ str_ends.c
@@ -4,8 +4,8 @@
 
 int str_ends(const char *s, const char *end)
 {
-  register unsigned int slen;
-  register unsigned int elen;
+  register unsigned long slen;
+  register unsigned long elen;
   slen = str_len(s);
   elen = str_len(end);
   if (elen > slen) elen = slen;
