+++ str_ndiff.c
@@ -1,6 +1,6 @@
 #include "str.h"
 
-int str_ndiff(const char *s, const char *t, unsigned int len)
+int str_ndiff(const char *s, const char *t, unsigned long len)
 {
   register char u;
 
@@ -11,6 +11,6 @@ int str_ndiff(const char *s, const char *t, unsigned int len)
     if (!len--) return 0; u = *s; if (u != *t) break; if (!u) break; ++s; ++t;
   }
 
-  return ((int)(unsigned int)(unsigned char) u)
-       - ((int)(unsigned int)(unsigned char) *t);
+  return ((int)(unsigned long)(unsigned char) u)
+       - ((int)(unsigned long)(unsigned char) *t);
 }
